(Backport) Curcial AfxHookSource fixes
These fixes should be backported to the main branch if required. I.e. there is
a mistake in asmClassTools.h with the class pointer (wrong offset), which
will cause trouble in cases where those JMP_..._OFS... functions are used
with a non-zero offset (I.e. can cause crashes on CAfxMesh::IndexCount).
Other fixes are preventing multiple hooking of a mesh / call queue aka accidental chaining.

Sadly I could not solve issue ripieces/advancedfx#22 yet, it seems to be a
game problem, when a lot is drawn, i.e. the skybox in that case multiple
times. Not sure if it's simply running out of reserved memory or if it's a
threading problem. Chances are this problem is in the game itself and not
HLAE related at all.